backend/requirements.txt
fastapi
uvicorn
pydantic
redis
psycopg2-binary
requests
PyJWT
numpy
python-dotenv
aiofiles

starlette
livekit
python-multipart


# ElevenLabs
elevenlabs==0.2.26

colorama==0.4.6

livekit-api